requirements.txt
sentence-transformers
chromadb
langchain
langchain-groq
streamlit
pandas
requests
python-dotenv
pydantic